Folding norm.impl.* into just norm package.
